add: MusicCommunityControllerAdvice 를 Music 도메인 전체에 해당하는 어드바이스로 확장 및 추가된 Exception들에 대한 Handling 로직 작성
